backend/requirements.txt
asgiref==3.11.1
black==25.11.0
click==8.1.8
coverage==7.10.7
coveralls==4.0.1
Django==4.2.28
django-dbml==0.10.1
django-mailjet==0.3.1
flake8==7.3.0
importlib_metadata==8.7.1
isort==6.1.0
mailjet-rest==1.4.0
mccabe==0.7.0
mypy_extensions==1.1.0
packaging==26.0
pathspec==1.0.4
platformdirs==4.4.0
psycopg2-binary==2.9.11
pycodestyle==2.14.0
pyflakes==3.4.0
pytokens==0.4.1
sqlparse==0.5.5
tblib==3.2.2
tomli==2.4.0
typing_extensions==4.15.0
zipp==3.23.0
drf-spectacular==0.29.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
cryptography==44.0.2
requests==2.32.3
Shapely>=2.0.0
pyproj>=3.6.0
pyshp==3.0.3
django-q2>=1.7.0
Pillow>=10.0.0
openpyxl>=3.1.0
